+++ Source/ilvyion.NutrientDispenserDispensableFix/ImprovedNutrientPasteDispenserModExtension.cs
@@ -1,0 +1,8 @@
+#nullable disable
+
+namespace ilvyion.NutrientDispenserDispensableFix;
+
+public class ImprovedNutrientPasteDispenserModExtension : DefModExtension
+{
+    public ThingDef dispensableDef;
+}